+++ lib/djvCore/Path.cpp
@@ -144,8 +144,8 @@ namespace djv
                 fseqFileNameParseSizes(in.c_str(), &sizes, FSEQ_STRING_LEN);
                 path = in.substr(0, sizes.path);
                 base = in.substr(sizes.path, sizes.base);
-                number = in.substr(sizes.path + sizes.base, sizes.number);
-                extension = in.substr(sizes.path + sizes.base + sizes.number, sizes.extension);
+                number = in.substr(sizes.path + static_cast<size_t>(sizes.base), sizes.number);
+                extension = in.substr(sizes.path + static_cast<size_t>(sizes.base) + static_cast<size_t>(sizes.number), sizes.extension);
             }
 
             std::vector<std::string> Path::splitDir(const std::string & value)
